+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/platform_webview_widget.dart
@@ -11,6 +11,13 @@ import 'webview_platform.dart';
 abstract class PlatformWebViewWidget extends PlatformInterface {
   /// Creates a new [PlatformWebViewWidget]
   factory PlatformWebViewWidget(PlatformWebViewWidgetCreationParams params) {
+    assert(
+      WebViewPlatform.instance != null,
+      'A platform implementation for `webview_flutter` has not been set. Please '
+      'ensure that an implementation of `WebViewPlatform` has been set to '
+      '`WebViewPlatform.instance` before use. For unit testing, '
+      '`WebViewPlatform.instance` can be set with your own test implementation.',
+    );
     final PlatformWebViewWidget webViewWidgetDelegate =
         WebViewPlatform.instance!.createPlatformWebViewWidget(params);
     PlatformInterface.verify(webViewWidgetDelegate, _token);