+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -3,6 +3,14 @@
 * Adds support for handling file selection. See `AndroidWebViewController.setOnShowFileSelector`.
 * Updates pigeon dev dependency to `4.2.14`.
 
+## 3.1.3
+
+* Fixes crash when the Java `InstanceManager` was used after plugin was removed from the engine.
+
+## 3.1.2
+
+* Fixes bug where an `AndroidWebViewController` couldn't be reused with a new `WebViewWidget`.
+
 ## 3.1.1
 
 * Fixes bug where a `AndroidNavigationDelegate` was required to load a request.